+++ src/input.c
@@ -1,0 +1,14 @@
+#include "input.h"
+
+char unixGetch(void) { // https://stackoverflow.com/questions/12710582/
+	char key;
+	struct termios oldt,newt;
+	tcgetattr(STDIN_FILENO, &oldt);
+	newt = oldt;
+	newt.c_lflag &= ~(ICANON | ECHO);
+	tcsetattr(STDIN_FILENO, TCSANOW, &newt);
+	key = getchar();
+	tcsetattr(STDIN_FILENO, TCSANOW, &oldt);
+	return key;
+}
+