+++ src/input.h
@@ -1,14 +1,12 @@
-#include <termios.h>
-#include <unistd.h>
+#ifndef INPUT_H
+#define INPUT_H
 
-int unixGetch(void) { // https://stackoverflow.com/questions/12710582/
-	struct termios oldt,newt;
-	int8_t character;
-	tcgetattr( STDIN_FILENO, &oldt );
-	newt = oldt;
-	newt.c_lflag &= ~( ICANON | ECHO );
-	tcsetattr( STDIN_FILENO, TCSANOW, &newt );
-	character = getchar();
-	tcsetattr( STDIN_FILENO, TCSANOW, &oldt );
-	return character; }
+#include <stdio.h>
+#include <stdint.h>
+#include <unistd.h>
+#include <termios.h>
+
+char unixGetch(void);
+
+#endif