+++ src/ui.h
@@ -1,0 +1,14 @@
+#ifndef UI_H
+#define UI_H
+
+#include <math.h>
+#include <stdio.h>
+#include <string.h>
+
+#define INVERT_COLORS "\033[7m\0"
+#define RESTORE_COLORS "\033[0m\0"
+
+void lowerBorder(void);
+
+#endif
+